refactor(web): extract pane controller glue into paneHostController (TASK-2170)
PLAN-2154 Phase 2 (Architecture E) increment 1 of 2. Move the pane-
navigation controller — the depth/ownership state machine, the fenced
`history.go` traversals + their `afterNavigate` settle, and the three-way
staged close — out of the 4805-line collection route into a shared
`$lib/collections/paneHostController` factory, so the collection page and
(in TASK-2174) the full-page item host mount ONE controller with no
duplicated arithmetic.

The pure decision logic already lived in `paneController.ts` (unit-tested);
this extracts the SvelteKit-bound glue that executes those plans. The five
collection-specific hooks are injected: getOpenItemRef, cancelFollow
(j/k pane-follow), focusPaneRegion (owned by the shell), captureReturnFocus
(list-row focus return), and setBypassNavGuard (unsaved-draft leave guard).

The route destructures the controller with the SAME names its markup,
handlers, and test hook already use, so no call sites changed; the depth-
aware ESC pop now routes through the public `handlePaneBack` primitive
instead of the now-internal `paneHistoryGo`.

Behavior-preserving: `npm run check` 0 errors, vitest 435 passed, pane e2e
49/49 (matches the pre-extraction baseline). Increment 2 extracts the
<aside> shell + resize/focus into PaneHost.svelte.
